requirements.txt
pandas>=1.3
pyarrow>=6.0.0
fastparquet>=0.8.0
requests>=2.25
pyyaml>=5.4
pytest>=6.2
python-dateutil>=2.8.2
